requirements.txt
pyyaml
argparse
nats.py
openai